+++ src/components/result/speechCautionSection/SpeechCautionSection.tsx
@@ -1,0 +1,20 @@
+import NavigateButton from 'components/common/NavigateButton';
+
+import CautionTitle from './CautionTitle';
+import CautionList from './CautionList';
+
+const HOME_BUTTON_TEXT = '홈으로 가기';
+
+const SpeechCautionSection = () => {
+  return (
+    <div className="px-6 pb-10 pt-[62px] bg-white">
+      <CautionTitle />
+      <CautionList />
+      <NavigateButton path="/" ga="result_home">
+        {HOME_BUTTON_TEXT}
+      </NavigateButton>
+    </div>
+  );
+};
+
+export default SpeechCautionSection;
